fix all register controllers bugs

+++ src/utils/cloudinary.js
@@ -1,4 +1,5 @@
 import { v2 as cloudinary } from 'cloudinary';
+import { log } from 'console';
 import fs from 'fs';
 
 cloudinary.config({ 
@@ -13,12 +14,16 @@ const uploadOnCloudinary = async(localFilePath) => {
         const result = await cloudinary.uploader.upload(localFilePath, {
             resource_type: "auto",
         });
-        console.log("file is uploaded on cloudinary",result.url);
-        return response;
+        // console.log("file is uploaded on cloudinary",result.url);
+        fs.unlinkSync(localFilePath);
+        return result;
         
     }
     catch(error){
-        fs.unlinkSync(localFilePath);
+        console.log("Error uploading to Cloudinary:", error);
+        if(fs.existsSync(localFilePath)){
+            fs.unlinkSync(localFilePath);
+        }
     }
 }
